requirements.txt
flask
python-docx
pydantic
python-dotenv
langchain
langchain-groq